remove default registry and add unittest
