--- python/oneflow/framework/infer_compiler_registry/register_diffusers/__init__.py
@@ -1,35 +1,0 @@
-"""
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-from oneflow.framework.infer_compiler.transform import register
-
-# register(package_names=["diffusers"])
-from diffusers.models.attention_processor import Attention, AttnProcessor2_0
-from diffusers.models.attention_processor import LoRAAttnProcessor2_0
-from diffusers.models.transformer_2d import Transformer2DModel
-from .attention_processor_oflow import Attention as AttentionOflow
-from .attention_processor_oflow import AttnProcessor as AttnProcessorOflow
-from .attention_processor_oflow import LoRAAttnProcessor2_0 as LoRAAttnProcessorOflow
-from .transformer_2d_oflow import Transformer2DModel as Transformer2DModelOflow
-
-torch2oflow_class_map = {
-    Attention: AttentionOflow,
-    AttnProcessor2_0: AttnProcessorOflow,
-    LoRAAttnProcessor2_0: LoRAAttnProcessorOflow,
-}
-if Transformer2DModelOflow is not None:
-    torch2oflow_class_map.update({Transformer2DModel: Transformer2DModelOflow})
-
-register(torch2oflow_class_map=torch2oflow_class_map)
