--- python/oneflow/framework/infer_compiler_registry/register_diffusers_quant/__init__.py
@@ -1,34 +1,0 @@
-"""
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-from infer_compiler import register
-
-import oneflow as flow
-import diffusers_quant
-
-torch2oflow_class_map = {
-    diffusers_quant.FakeQuantModule: diffusers_quant.OneFlowFakeQuantModule,
-    diffusers_quant.StaticQuantConvModule: diffusers_quant.OneFlowStaticQuantConvModule,
-    diffusers_quant.DynamicQuantConvModule: diffusers_quant.OneFlowDynamicQuantConvModule,
-    diffusers_quant.StaticQuantLinearModule: diffusers_quant.OneFlowStaticQuantLinearModule,
-    diffusers_quant.DynamicQuantLinearModule: diffusers_quant.OneFlowDynamicLinearQuantModule,
-}
-
-
-def convert_func(mod: flow.Tensor, verbose=False):
-    return mod
-
-
-register(torch2oflow_class_map=torch2oflow_class_map, torch2oflow_funcs=[convert_func])
